# aliancapanorama-src/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PAP - Projeto Aliança Panorama API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        tier:
          type: integer
          description: "0=guest 1=aluno1 2=aluno2 3=aluno3 4=aluno4 5=root/dev"
        displayName:
          type: ["string", "null"]
        subscriptionStatus:
          type: ["string", "null"]
          description: "active | cancelled | expired | suspended | past_due | null"
        lastDowngradeAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - login
        - tier
        - displayName
...